make

+++ app/controllers/books_controller.rb
@@ -3,16 +3,21 @@ class BooksController < ApplicationController
   def index
    @books= Book.all
    @book =Book.new
-  
+  @user= User.find(current_user.id)
   
   end
   
 
   def create
    @book= Book.new(post_book_params)
-   if @book.user_id= current_user.id
-   @book.save
-   redirect_to user_path(current_user.id),notice:"Book was successfully created."
+   if @book.user_id= current_user.id 
+      if@book.save
+  
+       redirect_to book_path(current_user.id),notice:"Book was successfully created."
+       else
+         redirect_to books_path
+         
+      end
    else
      render :index
      
@@ -20,14 +25,15 @@ class BooksController < ApplicationController
   end
 
   def show
-    @book= Book.find(params[:id])
-    @users= User.find(@book.user_id)
+    @book=Book.new
+    @books= Book.find(params[:id])
+    @users= User.find(@books.user_id)
 
   end
   
   def edit
     @book= Book.find(params[:id])
-    if @book.user.id == cureent_user
+    if @book.user.id == current_user.id
         render :edit
     else
         redirect_to user_path(@book.user.id)
@@ -44,9 +50,11 @@ class BooksController < ApplicationController
    end
   end   
   def destroy
-    book=Book.destroy
+    book=Book.find(params[:id])
+   if book.user_id == current_user.id
     book.destroy
     redirect_to books_path
+   end
   end
 
 
@@ -56,5 +64,10 @@ def post_book_params
   params.require(:book).permit(:title,:body,:user_id)
 
 end
+
+def user_params
+   params.require(:user).permit(:name,:profile_image,:introduction)
+   
+ end
 
 end
